+++ mixxx/lib/soundtouch-1.6.0/TDStretch.cpp
@@ -721,7 +721,7 @@ void TDStretch::acceptNewOverlapLength(int newOverlapLength)
 
         pRefMidBufferUnaligned = new SAMPLETYPE[2 * overlapLength + 16 / sizeof(SAMPLETYPE)];
         // ensure that 'pRefMidBuffer' is aligned to 16 byte boundary for efficiency
-        pRefMidBuffer = (SAMPLETYPE *)((((ulong)pRefMidBufferUnaligned) + 15) & (ulong)-16);
+        pRefMidBuffer = (SAMPLETYPE *)((((uintptr_t)pRefMidBufferUnaligned) + 15) & (uintptr_t)-16);
     }
 }
